HTML error fixed
It's an HTML rule violation error bro, there can't be another <a> tag between the <a> </a> tag, I searched index.js because it is on the first page. If <Button as="a" /> with the "as" attribute tries to render the <Button> component with the <a> tag, this will break the HTML rule. Because the button <button> and link <a> tags are designed for different purposes and are not interchangeable.